requirements.txt
Django==5.0.6
mysqlclient>=2.2.1
django-environ>=0.11

PyMySQL>=1.1
mysql-connector-python>=8.0
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27

djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
gunicorn==22.0.0

pycryptodome

faiss-cpu
google-genai
numpy
google-generativeai